requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Django==4.0.2
django-crispy-forms==1.14.0
fabio==0.13.0
favicon==0.7.0
fontawesomefree==6.0.0
fonttools==4.29.1
h5py==3.6.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
requests==2.27.1
silx==1.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==2.67.0
tzdata==2021.5
urllib3==1.26.8
